+++ classes/book.rb
@@ -3,11 +3,11 @@ require_relative 'item'
 class Book < Item
   attr_reader :publisher, :cover_state, :title
 
-  def initialize(publish_date, publisher, cover_state, title)
+  def initialize(title, publish_date, publisher, cover_state)
     super(publish_date)
+    @title = title
     @publisher = publisher
     @cover_state = cover_state
-    @title = title
   end
 
   def can_be_archived?
