--- data/books.json
@@ -1,1 +1,0 @@
-[{"id":906,"title":"War and Peace","publisher":"RB Traders","publish_date":"2023-02-15","cover_state":"GOOD"},{"id":910,"title":"Hello Microverse","publisher":"David Igbo","publish_date":"2022-01-01","cover_state":"GOOD"},{"id":361,"title":"The Ruby ","publisher":"Christian Roy","publish_date":"1990-01-02","cover_state":"BAD"}]
