--- data/labels.json
@@ -1,1 +1,0 @@
-[{"id":997,"title":"War and Peace","color":"Red"},{"id":494,"title":"Hello Microverse","color":"Green"},{"id":44,"title":"The Ruby ","color":"Purple"}]
