+++ music_album.json
@@ -34,5 +34,11 @@
     "publish_date": "2012-12-12",
     "on_spotify": true,
     "genre": "ds"
+  },
+  {
+    "name": "adsf",
+    "publish_date": "2010-10-10",
+    "on_spotify": true,
+    "genre": "asd"
   }
 ]
